Export generateGuid function to src/utils folder
