+++ onboarding-task/src/utils/utils.js
@@ -1,0 +1,13 @@
+
+function _generateGuid() {
+  function s4() {
+    return Math.floor((1 + Math.random()) * 0x10000)
+      .toString(16)
+      .substring(1);
+  }
+  return s4() + s4() + '-' + s4() + '-' + s4() + '-' +
+    s4() + '-' + s4() + s4() + s4();
+}
+
+export { _generateGuid };
+
